classic_automation/requirements.txt
pyautogui>=0.9.54
pygetwindow>=0.0.9
Pillow>=10.0.0
psutil>=5.9.0
sympy>=1.12
python-docx>=1.1.0
mcp>=1.0.0
